Addressed the remarks (move test to a standalone class, change from iterator to index traversal,  keep last processing time)
